Reorganise existing code into organised module folder
